scripts: dependencies: do not specify any version
Allow for more flexibility when it comes to dependencies. The scripts do
not have any strong dependency or known compatiblity issues, so do not
specify any version. This means the latest version will be downloaded.
If anything comes up scripts can always be updated or versions until
scripts are updated.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>