scripts/genpinctrl: Add remap variant if any
Pin remap configurations may take two different values
For instance:
 USART3_CTS PA13 matches both remap values 00 and 01.

In a peripheral pin remapping configuration, all pins associated to the
device must share the same remap value.
For instance, on USART3
-remap0 cfg uses PB10, PB11, PB12, PB13, PB14, all configured with remap0
-remap1 cfg uses PC10, PC11, PC12, PB13, PB14, all configured with remap1

If pinctrl driver detects one of the pins has a remap value different
from the whole set, it will generate an error.

As a consequence, it may happen that for a single pin (PB13, in the above
example), we need to generate both remap configurations, each one
correctly identified with a specific name.

To take this into account, always include the remap information in
the generated pin configuration node.
T
o minimize the impact, it has been decided to apply this only if remap
value differs for "0" or "No remap".

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>