lib: hci: Add stm32wba BLE controller library
Make required changes to port STM32WBA BLE controller library on Zephyr.

This work is made of two parts:
- Configure zephyr/module.yaml to allow importation of the 2 required
binary blobs that are required to get a BLE stack working with STM32WBA
A copy of the license required to use these binary blobs (ST ULTIMATE
LIBERTY) is made available as part of this commit
- Import bluetooth controller related files that are needed to implement
a BLE controller library. These files are provided with OSI approved
Licenses (BSD-3 of MIT depending on the case).

Signed-off-by: Erwan Gouriou <erwan.gouriou@st.com>