[DNM] dts: remove duplicate pinctrl nodes
This patch cannot be merged as-is; I'm just using it for now to get
cleaner CI in https://github.com/zephyrproject-rtos/zephyr/pull/50287.

The STM32 G0 and H7 family pinctrl DTSI files have a lot of duplicate
nodes. In most cases, they are simply redundant, but not always, and I
suspect there may be bugs in the genpinctrl script or the open pin
database.

For example, look at the usb_otg_hs_ulpi_dir_pc2 nodes in
stm32h735igkx-pinctrl.dtsi:

  /omit-if-no-ref/ usb_otg_hs_ulpi_dir_pc2: usb_otg_hs_ulpi_dir_pc2 {
          pinmux = <STM32_PINMUX('C', 2, ANALOG)>;
  };

  /omit-if-no-ref/ usb_otg_hs_ulpi_dir_pc2: usb_otg_hs_ulpi_dir_pc2 {
          pinmux = <STM32_PINMUX('C', 2, AF10)>;
  };

Note the different alternate function -- ANALOG vs. AF10.

With current dtlib, AF10 wins, but this will be an error once zephyr
PR 50287 is merged, and it looks suspicious.

The generation of these files should be investigated and fixed to not
generate duplicate or conflicting definitions by someone who knows the
hardware better; this is just a demo of what sort of thing needs to
happen.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>