scripts: ble_library: Simplify script
Remove usage of template for CMakeLists.txt and README generation when
updating BLE library.

Generating CMakeLists.txt for BLE lib is not something that has added
value, in the contrary of CMakeLists.txt use in Cube Package, for which
each file has similar usage. In BLE lib, each file is specific and this
is a developper choice if a file should be added or removed.
Besides, each time this file is generated, the file list ordering is
modified with no reason, which is only generating noise.

Similarly, fully generating README in current case is error prone
as user can't suspect this is a generated file. Hence, users tend
to update README file directly rather than the template file itself.
Modify this to only update the SHA1 and version information.

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>