scripts: genpinctrl: Generate SWJ port pins
Generate Serial Wire and JTAG port pins definitions.
The particularity of these signals is that they are part of
default SoC boot time configuration (in order to enable debug at boot time
with no SW configuration required).

The reason we need to define them is to enable the ability to set
JTAG port pins in analog mode when they are not needed in order to
save power (set them to analog).

Today, the use case of dynamic configuration from analog back to a
functional SWJ state is not taken into account, hence we're not generating
the complete pin configuration (including possible pull-down or slew-rate
combinations).

Similarly, we're not generating these signals on STM32F1 as default SoC
configuration is analog state. Hence no need to be taken into account
for the required functionality.

Signed-off-by: Erwan Gouriou <erwan.gouriou@st.com>