stm32cube: Remove legacy headers inclusion
These headers provide redirections from legacy symbols definitions.
They are provided only for backward compatibility, but in context of Zephyr
integration we're supposed to update stm32 drivers to stay up to date
with latest versions ath using deprecated symbols.

Purely remove the inclusion and update symbols used in Zephyr if required.

As a benefit this should minimize the possibility of conflicts with Zephyr
or other subsystems API (such as POSIX).

Signed-off-by: Erwan Gouriou <erwan.gouriou@st.com>