stm32cube: Remove all STM32Cube release notes
Remove all STM32Cube release notes since the links to release notes of
all components are relative paths to Cube package.
So, they are broken links.
In the README file of the each cube folder, there is a link to
the GitHub containing the release notes and all the STM32Cube details.

Signed-off-by: Nidhal BEN OTHMEN <nidhal.benothmen@st.com>