stm32cube: stm32mp2xx: eth: Fix ETH_DMA_XXX definitions
On the Ethernet peripheral, there are two DMA channels.
As a consequence, the ETH_DMACSR_XXX definitions do not exist and
should be replaced by ETH_DMAC0SR_CDE and ETH_DMAC1SR_CDE.

While waiting for the fix in the STM32Cube distribution, the workaround
is to replace ETH_DMACSR_XXX with ETH_DMAC0SR_XXX.
Indeed, the bit definitions for ETH_DMAC0SR_XXX and ETH_DMAC1SR_XXX
are identical.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>