stm32cube: remove HAL Legacy definitions header from build files
Remove the HAL Legacy definitions header from all series using a little
trick to ensure proper compilation despite the missing file, and update
the README of all series to remove mention of patches to the header we are
getting rid of.

Also remove the other legacy headers (can_legacy / eth_legacy) which are
never used from the Zephyr side. This is however more of a cosmectic
patch as these headers are not part of the build (they are only included
if the corresponding HAL module is enabled, which is never done in Zephyr).

Signed-off-by: Mathieu Choplain <mathieu.choplain-ext@st.com>